Fix compile warnings from gcc and clang.
There are still warnings in lodepng_unittest.cpp which are being
suppressed by a compile option. They should be fixed. See comment
in CMakeLists.txt.

+++ examples/example_sdl.cpp
@@ -45,9 +45,9 @@ int show(const std::string& caption, const unsigned char* rgba, unsigned w, unsi
   if(w / 1024 >= jump) jump = w / 1024 + 1;
   if(h / 1024 >= jump) jump = h / 1024 + 1;
 
-  size_t screenw = w / jump;
-  size_t screenh = h / jump;
-  size_t pitch = screenw * sizeof(Uint32);
+  int screenw = w / jump;
+  int screenh = h / jump;
+  int pitch = screenw * sizeof(Uint32);
   //init SDL
   SDL_Window* sdl_window;
   SDL_Renderer* sdl_renderer;
